// pbweberogwcl/src/app/recuperi/commons/models/dettaglio-recupero.ts
/*
 * Copyright Regione Piemonte - 2024 
 * SPDX-License-Identifier: EUPL-1.2 
*/

/* tslint:disable */
export interface DettaglioRecupero {
  dataRecupero?: string;
  estremiDeterminaRecupero?: string;
  idAnnoContabile?: number;
  idRecupero?: number;
  idTipoRecupero?: number;
  importoAgevolato?: number;
  importoErogato?: number;
  importoRecuperato?: number;
  importoRecupero?: number;
  importoRevocato?: number;
  modalitaAgevolazione?: string;
  noteRecupero?: string;
}